add messages

+++ server/src/main/java/com/developmentontheedge/be5/operations/DeleteOperation.java
@@ -3,6 +3,7 @@ package com.developmentontheedge.be5.operations;
 import com.developmentontheedge.be5.metadata.model.TableReference;
 import com.developmentontheedge.be5.operation.OperationResult;
 import com.developmentontheedge.be5.operation.OperationSupport;
+import com.developmentontheedge.be5.operation.TransactionalOperation;
 import com.developmentontheedge.be5.util.Utils;
 
 import java.util.List;
@@ -12,7 +13,7 @@ import static com.developmentontheedge.be5.metadata.model.EntityType.COLLECTION;
 import static com.developmentontheedge.be5.metadata.model.EntityType.GENERIC_COLLECTION;
 
 
-public class DeleteOperation extends OperationSupport
+public class DeleteOperation extends OperationSupport implements TransactionalOperation
 {
     @Override
     public Object getParameters(Map<String, Object> presetValues) throws Exception
@@ -23,26 +24,46 @@ public class DeleteOperation extends OperationSupport
     @Override
     public void invoke(Object parameters) throws Exception
     {
-        database.getEntity(getInfo().getEntityName()).remove(context.records);
+        StringBuilder out = new StringBuilder();
+
+        int updateCount = database.getEntity(getInfo().getEntityName()).remove(context.records);
+
+        out.append( "" + updateCount + " " + ( "records were deleted from" ) + " <i>" + getInfo().getEntityName() + "</i><br />" );
 
         List<TableReference> collectionRefs = meta.getRefToTable(COLLECTION, getInfo().getEntityName());
 
         for (TableReference reference : collectionRefs)
         {
-            db.update("DELETE FROM " + reference.getTableFrom() +
+            int updateCount1 = db.update("DELETE FROM " + reference.getTableFrom() +
                             " WHERE " + reference.getColumnsFrom() + " IN " + Utils.inClause(context.records.length),
                     (Object[]) context.records);
+
+            if( updateCount1 > 0 )
+            {
+                //todo localizedMessage
+                out.append( "" + updateCount1 +
+                        " " + ( "records were deleted from the collection" ) + " <i>" + reference.getTableFrom() + "</i><br />" );
+            }
         }
 
-        List<TableReference> genericCollectionRefs = meta.getRefToTable(GENERIC_COLLECTION, getInfo().getEntityName());
-
-        for (TableReference reference : genericCollectionRefs)
+        if( !GENERIC_COLLECTION.equals( getInfo().getEntity().getType() ) )
         {
-            db.update("DELETE FROM " + reference.getTableFrom() +
-                            " WHERE " + reference.getColumnsFrom() + " IN " + Utils.inClause(context.records.length),
-                    (Object[]) Utils.addPrefix(context.records, getInfo().getEntityName()));
+            List<TableReference> genericCollectionRefs = meta.getRefToTable(GENERIC_COLLECTION, getInfo().getEntityName());
+
+            for (TableReference reference : genericCollectionRefs)
+            {
+                int updateCount1 = db.update("DELETE FROM " + reference.getTableFrom() +
+                                " WHERE " + reference.getColumnsFrom() + " IN " + Utils.inClause(context.records.length),
+                        (Object[]) Utils.addPrefix(context.records, getInfo().getEntityName()));
+
+                if( updateCount1 > 0 )
+                {
+                    out.append( "" + updateCount1 +
+                            " " + ( "records were deleted from the generic collection" ) + " <i>" + reference.getTableFrom() + "</i><br />" );
+                }
+            }
         }
 
-        setResult(OperationResult.finished());
+        setResult(OperationResult.finished(out.toString()));
     }
 }
